Refractored Model class
